Refactored using helpers
