+++ lambdas/login/helpers/google-auth/process-error.js
@@ -1,0 +1,6 @@
+module.exports = response => err => {
+  console.log(JSON.stringify(err, null, 2))
+  response.statusCode = 401
+  response.body = JSON.stringify({ message: 'Provided credentials are incorrect.' })
+  return response
+}
