+++ lambdas/login/helpers/google-auth/process-response.js
@@ -1,0 +1,4 @@
+module.exports = response => res => {
+  response.body = JSON.stringify({ idToken: res.sessionToken })
+  return response
+}
